[CodeQL] XmlSupportPkg: Fixing CodeQL failures in this package (#173)
# Preface

Please ensure you have read the [contribution
docs](https://github.com/microsoft/mu/blob/master/CONTRIBUTING.md) prior
to submitting the pull request. In particular,
[pull request
guidelines](https://github.com/microsoft/mu/blob/master/CONTRIBUTING.md#pull-request-best-practices).

## Description

Pointer overflow has undefined behavior according to the C and C++
standards.
This change will convert the pointer to `UINTN` prior to the comparison
and leverage
unsigned integer overflow, which is defined, to perform the overflow
check.

There is also an error involving variable comparison between different
width.

For each item, place an "x" in between `[` and `]` if true. Example:
`[x]`.
_(you can also check items in the GitHub UI)_

- [x] Impacts functionality?
- **Functionality** - Does the change ultimately impact how firmware
functions?
- Examples: Add a new library, publish a new PPI, update an algorithm,
...
- [x] Impacts security?
- **Security** - Does the change have a direct security impact on an
application,
    flow, or firmware?
  - Examples: Crypto algorithm change, buffer overflow fix, parameter
    validation improvement, ...
- [ ] Breaking change?
- **Breaking change** - Will anyone consuming this change experience a
break
    in build or boot behavior?
- Examples: Add a new library class, move a module to a different repo,
call
    a function in a new library class in a pre-existing module, ...
- [ ] Includes tests?
  - **Tests** - Does the change include any explicit test code?
  - Examples: Unit tests, integration tests, robot tests, ...
- [ ] Includes documentation?
- **Documentation** - Does the change contain explicit documentation
additions
    outside direct code modifications (and comments)?
- Examples: Update readme file, add feature readme file, link to
documentation
    on an a separate Web page, ...

## How This Was Tested

Verified functional on Q35 virtual platform.

## Integration Instructions

N/A

---------

+++ XmlSupportPkg/Library/XmlTreeLib/XmlTreeLib.c
@@ -996,7 +996,7 @@ BuildNodeList (
       //
       // See if we have a value.
       //
-      for (UINT32 i = 0; i < Next.Run.ulCharacters; i++) {
+      for (UINT64 i = 0; i < Next.Run.ulCharacters; i++) {
         if (!IsWhiteSpace (LocalHyperSpace[i])) {
           // MS_CHANGE
           NotWhiteSpace = TRUE;
